+++ packages/ketcher-core/src/application/render/restruct/reatom.ts
@@ -34,7 +34,7 @@ import {
 } from './generalEnumTypes';
 
 import ReObject from './reobject';
-import ReStruct from './restruct';
+import type ReStruct from './restruct';
 import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
 import draw from '../draw';
