+++ packages/ketcher-core/src/application/render/restruct/reenhancedFlag.ts
@@ -18,7 +18,7 @@ import { Box2Abs, Fragment, StereoFlag } from 'domain/entities';
 
 import { LayerMap } from './generalEnumTypes';
 import ReObject from './reobject';
-import ReStruct from './restruct';
+import type ReStruct from './restruct';
 import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
