+++ packages/ketcher-core/src/application/render/restruct/reobject.ts
@@ -15,7 +15,7 @@
  ***************************************************************************/
 
 import { Box2Abs } from 'domain/entities';
-import ReStruct from './restruct';
+import type ReStruct from './restruct';
 import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
 import Visel from './visel';
