+++ packages/ketcher-core/src/application/render/restruct/restruct.ts
@@ -40,7 +40,7 @@ import ReRxnPlus from './rerxnplus';
 import ReSGroup from './resgroup';
 import ReSimpleObject from './resimpleObject';
 import ReText from './retext';
-import { Render } from '../raphaelRender';
+import type { Render } from '../raphaelRender';
 import Visel from './visel';
 import util from '../util';
 import { ReRGroupAttachmentPoint } from './rergroupAttachmentPoint';
