+++ packages/ketcher-core/src/application/render/restruct/retext.ts
@@ -25,7 +25,7 @@ import { flatten, isEqual } from 'lodash/fp';
 
 import { LayerMap } from './generalEnumTypes';
 import ReObject from './reobject';
-import ReStruct from './restruct';
+import type ReStruct from './restruct';
 import { Scale } from 'domain/helpers';
 import { RaphaelBaseElement } from 'raphael';
